requirements.txt
google-auth==2.37.0
google-api-python-client==2.163.0
requests==2.32.3
Pillow==11.1.0
beautifulsoup4==4.12.3
openai==1.59.9
PyNaCl==1.5.0
